Tracked down weird rendering difference...
Updating Juniper to 0.1.9 only changed the Ziggurat class, for generating normal-distributed numbers. This is used by RotationTools in SquidSquad, which is used heavily by CyclicNoise, which is used by the swirl effect in Isonomicon. So when the swirl effect changed, I figured it was probably a seed change, but I also went through and debugged it more thoroughly so hopefully this should be less likely in the future.