Try a custom color difference metric.
This halves L, making all colors closer on the lightness axis, but still equally distant on A and B. It also reduces the total distance returned by the metric, which may help. The dither strength is a little lower, but this doesn't have as much effect as the other changes.